+++ hass-actronque/hass-actronque/AirConditionerData.cs
@@ -10,6 +10,7 @@ namespace HMX.HASSActronQue
 		public string FanMode; // FanMode
 		public string Mode; // Mode
 		public bool On; // isOn
+		public bool CleanFilter; // Indoor Filter Time to Clean
 		public double SetTemperatureCooling; // TemperatureSetpoint_Cool_oC
 		public double SetTemperatureHeating; // TemperatureSetpoint_Heat_oC
 		public double Temperature; // LiveTemp_oC
@@ -21,6 +22,7 @@ namespace HMX.HASSActronQue
 		public double CoilInletTemperature; // CoilInlet
 		public double FanPWM; // FanPWM
 		public double FanRPM; // FanRPM
+		public double FanTSFC; // Fan Time Since Filter Cleaned - Hours
 		public DateTime LastUpdated;
 
 		public AirConditionerData()